+++ NEWS.md
@@ -1,3 +1,5 @@
+# connectapi (development version)
+
 # connectapi 0.6.0
 
 ## New features